Refactoring; split `tests/forms/test_client` test suite into `tests/forms/test_client_register` and `tests/forms/test_client_edit.py`
